pyproject.toml
[project]
name = "knowledge-kiwi"
version = "0.1.0"
description = "Knowledge base MCP server with Zettelkasten support"
authors = [{name = "Leo", email = "leo@example.com"}]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "mcp>=1.0.0,<2.0.0",
    "supabase>=2.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "markdown>=3.5.0",
    "pyyaml>=6.0.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
knowledge-kiwi = "knowledge_kiwi.cli:main"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
